Flag and unveil all surrounding fields
If the number of flags in a safe field's environment is equal to its
label, unveil all of them when clicked.

If the number of potential mines in a safe field's environment is equal
to its label, flag all of them.

This allows you to kind of play the game now.