Do not automatically re-toggle buttons
Using set_active() in the clicked() handler is not such a good idea
because it will re-run the handler.  This is actually what we wanted
because the second run then calls the Logic::pressed() function, but
that is hard to tell from the code, so we should not do it this way.

Instead, we should just call Logic::pressed() every time, and let it
handle the re-toggling (via set_field_state()).