Revert "Catch button release instead of press"
This reverts commit 8e02fba3e3a92837f99b3019d92de9c1b11a9f36.  Turns out
you actually cannot correct your fault because Gtk will always issue the
release event on the field where the button was pressed, independently
of where the cursor is now.

Until I found a way to work around this issue, just catch the button
press again so we at least get a low latency.