Remove GUI callbacks
Owning an object and passing self to it when calling its methods seems
rather unrust-y to me (and it is causing a bit of an issue when we want
the Logic module to work for hypothetical fields, that are not reflected
in the GUI), so change that whole thing.  Instead of using callbacks
provided by GUI, Logic will now create a list of updates that the UI can
request once the Logic method returns.